+++ experimental/ingestion/airbyte/terraform/variant-a-yamldecode/versions.tf
@@ -1,0 +1,9 @@
+terraform {
+  required_version = ">= 1.5"
+  required_providers {
+    airbyte = {
+      source  = "airbytehq/airbyte"
+      version = "1.2.0"
+    }
+  }
+}